+++ internal/provider/coffees_data_source.go
@@ -1,0 +1,35 @@
+package provider
+
+import (
+	"context"
+
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+)
+
+// Ensure the implementation satisfies the expected interfaces.
+var (
+	_ datasource.DataSource = &coffeesDataSource{}
+)
+
+// NewCoffeesDataSource is a helper function to simplify the provider implementation.
+func NewCoffeesDataSource() datasource.DataSource {
+	return &coffeesDataSource{}
+}
+
+// coffeesDataSource is the data source implementation.
+type coffeesDataSource struct{}
+
+// Metadata returns the data source type name.
+func (d *coffeesDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_coffees"
+}
+
+// Schema defines the schema for the data source.
+func (d *coffeesDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+	resp.Schema = schema.Schema{}
+}
+
+// Read refreshes the Terraform state with the latest data.
+func (d *coffeesDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+}
