+++ internal/provider/deprecated.coffees_data_source.go
@@ -1,25 +1,25 @@
-package provider
+// package provider
 
-import (
-	"context"
+// import (
+// 	"context"
 
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
-)
+// 	"github.com/hashicorp/terraform-plugin-framework/datasource"
+// 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+// )
 
-func NewCoffeesDataSource() datasource.DataSource {
-	return &coffeesDataSource{}
-}
+// func NewCoffeesDataSource() datasource.DataSource {
+// 	return &coffeesDataSource{}
+// }
 
-type coffeesDataSource struct{}
+// type coffeesDataSource struct{}
 
-func (d *coffeesDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_coffees"
-}
+// func (d *coffeesDataSource) Metadata(_ context.Context, req datasource.MetadataRequest, resp *datasource.MetadataResponse) {
+// 	resp.TypeName = req.ProviderTypeName + "_coffees"
+// }
 
-func (d *coffeesDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
-	resp.Schema = schema.Schema{}
-}
+// func (d *coffeesDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+// 	resp.Schema = schema.Schema{}
+// }
 
-func (d *coffeesDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
-}
+// func (d *coffeesDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
+// }
